Feat(1/autenticacion(registrar usuario)): Se terminaron de integrar las funcionalidades esteticas a la vissta de regitro de usuario.
Explicacion detallada:
- Se creo un componente para mejorar la mantenibilidad de los botones de envio con representacion de carga.
- Se reacomodo la vista regiter eparando logica de envio de estructura de vista e imagen
- Se corrigieron errores en el tipado
